[refactor] : alertError를 컴포넌트와 가깝게 위치
